added typescript types to get-user-code-frame file
